Remove old files, update configs, add gfal support options, rework TupleMaker
