+++ Studies/DNN/config/step2_training_setup_doubleLep.yaml
@@ -1,23 +1,9 @@
 UseParametric: true
-adv_activation: relu
-adv_grad_factor: 1.0
-adv_learning_rate: 0.001
-adv_model: false
-adv_submodule_steps: 50
-adv_submodule_tracker: 0
-adv_weight_decay: 0.004
 apply_common_gradients: true
 batch_compression_factor: 10
-nClasses: 4
-class_activation: tanh
-class_grad_factor: 0.1
-common_activation: tanh
-continue_model: DNN_Models/default/default.keras
-continue_training: false
+nClasses: 2
 disco_activation: tanh
-disco_lambda_factor: 100
-do_step2: false
-do_training: true
+disco_lambda_factor: 0
 dropout: 0.0
 features:
   - lep1_pt
@@ -40,16 +26,13 @@ highlevelfeatures:
   - dPhi_jet1_jet2
   - dPhi_MET_dilep
   - dPhi_MET_dibjet
-  # - min_dR_lep0_jets # For some reason, this is sometimes inf
-  # - min_dR_lep1_jets # Check event 2030682 in this dataset may30
   - MT
   - MT2_ll
   - MT2_bb
   - MT2_blbl
   - ll_mass
   - CosTheta_bb
-hmefeatures:
-  - DoubleLep_DeepHME_mass
+hmefeatures: []
 input_folder: /eos/user/d/daebi/DNN_Training_Datasets/DoubleLepton_v3/Dataset_Run3_2022/
 learning_rate: 0.001
 listfeatures:
@@ -93,38 +76,11 @@ listfeatures:
       - SelectedFatJet_SubJet2_phi
       - SelectedFatJet_SubJet2_pt
     - 0
-n_adv_layers: 3
-n_adv_units: 128
-n_class_layers: 3
-n_class_units: 128
-n_common_layers: 5
-n_common_units: 256
 n_disco_layers: 5
 n_disco_units: 512
 n_epochs: 40
 parametric_list:
-  - 250
-  - 260
-  - 270
-  - 280
-  - 300
-  - 350
-  - 450
-  - 550
   - 600
-  - 650
-  - 700
-  - 800
-  - 1000
-  - 1200
-  - 1400
-  - 1600
-  - 1800
-  - 2000
-  - 2500
-  - 3000
-  - 4000
-  - 5000
 patience: 100
 use_batch_norm: false
 weight_decay: 0.004
